requirement.txt
python-wordpress-xmlrpc==2.3
requests==2.13.0
rsa==3.4.2
sqlalchemy==1.1.9
pymysql==0.7.11
python-flickr==0.3.2
